# swagger.yaml
openapi: 3.0.0
info:
    title: 'API Documentation'
    version: 1.0.0
    description: 'API documentation for plans, workouts, and exercises.'
servers:
    -
        url: 'http://localhost:3010'
paths:
    /exercises:
        post:
            summary: 'Create a new exercise'
            tags: [Exercises]
            requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {name: {type: string, example: 'Push Up'}, sets: {type: integer, example: 3}, reps: {type: integer, example: 15}, restTime: {type: integer, example: 60}, tips: {type: string, example: 'Keep your back straight.'}, workout_id: {type: integer, example: 1}}}}}}
            responses: {'201': {description: 'Exercise created successfully', content: {application/json: {schema: {type: object, properties: {id: {type: integer, example: 1}, name: {type: string, example: 'Push Up'}, sets: {type: integer, example: 3}, reps: {type: integer, example: 15}, restTime: {type: integer, example: 60}, tips: {type: string, example: 'Keep your back straight.'}, workout_id: {type: integer, example: 1}}}}}}, '400': {description: 'All fields are required'}, '403': {description: 'Access denied. You do not have permission to create this exercise.'}, '422': {description: 'Invalid input values'}}
        get:
            summary: 'Get all exercises'
            tags: [Exercises]
            responses: {'200': {description: 'A list of exercises', content: {application/json: {schema: {type: array, items: {type: object, properties: {id: {type: integer, example: 1}, name: {type: string, example: 'Push Up'}, sets: {type: integer, example: 3}, reps: {type: integer, example: 15}, restTime: {type: integer, example: 60}, tips: {type: string, example: 'Keep your back straight.'}, workout_id: {type: integer, example: 1}}}}}}}, '403': {description: 'Access denied. You do not have the required permissions.'}, '500': {description: 'Internal server error'}}
    '/exercises/{exercise_id}':
        get:
            summary: 'Get a specific exercise'
            tags: [Exercises]
            parameters: [{name: exercise_id, in: path, required: true, description: 'ID of the exercise', schema: {type: integer}}]
            responses: {'200': {description: 'Exercise found', content: {application/json: {schema: {type: object, properties: {id: {type: integer, example: 1}, name: {type: string, example: 'Push Up'}, sets: {type: integer, example: 3}, reps: {type: integer, example: 15}, restTime: {type: integer, example: 60}, tips: {type: string, example: 'Keep your back straight.'}, workout_id: {type: integer, example: 1}}}}}}, '403': {description: 'Access denied. You do not have the required permissions.'}, '404': {description: 'Exercise not found'}}
        put:
            summary: 'Update an exercise'
            tags: [Exercises]
            parameters: [{name: exercise_id, in: path, required: true, description: 'ID of the exercise', schema: {type: integer}}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {name: {type: string, example: 'Push Up'}, sets: {type: integer, example: 3}, reps: {type: integer, example: 15}, restTime: {type: integer, example: 60}, tips: {type: string, example: 'Keep your back straight.'}}}}}}
            responses: {'200': {description: 'Exercise updated successfully'}, '400': {description: 'All fields are required'}, '403': {description: 'Access denied. You are not the owner of the plan or an admin.'}, '404': {description: 'Exercise not found'}, '422': {description: 'Invalid input values'}}
        delete:
            summary: 'Delete an exercise'
            tags: [Exercises]
            parameters: [{name: exercise_id, in: path, required: true, description: 'ID of the exercise', schema: {type: integer}}]
            responses: {'200': {description: 'Exercise deleted successfully'}, '403': {description: 'Access denied. You are not the owner of the plan or an admin.'}, '404': {description: 'Exercise not found'}}
...